refactor(map): split visualization shared color and legend modules
Extract color palette logic and resource legend glyph helpers out of tile_map_visualization_shared to reduce mixed concerns while keeping compatibility through shared re-exports. Refs #2322.
